chore: refactor account footer into separate AccountLogoutButton component for improved structure
