+++ src/app/(auth)/account/page.tsx
@@ -1,12 +1,11 @@
 import type { Metadata } from 'next'
 import { Suspense } from 'react'
 import { createMetadata } from '@/lib/metadata'
-import { AccountFooter } from './_components/account-footer'
+import { AccountLogoutButton } from './_components/account-logout-button'
 import { AccountSettingsSection } from './_components/account-settings-section'
 import { AccountSettingsSkeleton } from './_components/account-settings-skeleton'
 import { ActiveSessionsSection } from './_components/active-sessions-section'
 import { ActiveSessionsSkeleton } from './_components/active-sessions-skeleton'
-import { requireSession } from './_lib/session'
 
 export async function generateMetadata(): Promise<Metadata> {
   return createMetadata({
@@ -21,19 +20,20 @@ export async function generateMetadata(): Promise<Metadata> {
   })
 }
 
-export default async function AccountPage() {
-  const session = await requireSession()
-
+export default function AccountPage() {
   return (
     <div className='flex min-h-full w-full flex-1 flex-col p-4 md:p-6'>
       <main className='flex flex-1 flex-col gap-8'>
-        <header className='space-y-2'>
-          <h1 className='typography-title font-medium text-2xl'>
-            Account Settings
-          </h1>
-          <p className='typography-body text-muted-foreground text-sm'>
-            Manage your account information.
-          </p>
+        <header className='flex flex-col gap-4 sm:flex-row sm:items-start sm:justify-between'>
+          <div className='space-y-2'>
+            <h1 className='typography-title font-medium text-2xl'>
+              Account Settings
+            </h1>
+            <p className='typography-body text-muted-foreground text-sm'>
+              Manage your account information.
+            </p>
+          </div>
+          <AccountLogoutButton />
         </header>
         <Suspense fallback={<AccountSettingsSkeleton />}>
           <AccountSettingsSection />
@@ -45,7 +45,6 @@ export default async function AccountPage() {
           </Suspense>
         </div>
       </main>
-      <AccountFooter email={session.user.email} />
     </div>
   )
 }
